Added OpenCV.
Right now I'm just using the pip wheel since it's pretty recent (3.4.4). If this becomes too outdated later on, we can make it ourselves from source.

I chose to not make it since I'm busy and short on time and because it's already a decently new version. If anyone wants to build it themselves please feel free to add it in.